+++ 502/vite-project/src/App.tsx
@@ -2,7 +2,6 @@ import { useState } from 'react'
 import reactLogo from './assets/react.svg'
 import viteLogo from '/vite.svg'
 import './App.css'
-import './class1/A01025119/dokyu_main.tsx';
 
 function App() {
   const [count, setCount] = useState(0)
@@ -26,7 +25,7 @@ function App() {
           Edit <code>src/App.tsx</code> and save to test HMR
         </p>
         <p>
-          <a href="/A01025119" target="_blank" rel="noopener noreferrer">
+          <a href="/src/class1/A01025119/webpage/A01025119.html" target="_blank" rel="noopener noreferrer" >
             Pagina de Do Kyu
           </a>
         </p>
