+++ 502/vite-project/src/class1/A01025119/Tarea/Destruction.tsx
