+++ 502/vite-project/src/class1/A01025119/webpage/A01025119.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8" />
+  <link rel="stylesheet" href="/main_design.css" />
+  <link rel="stylesheet" href="/App.css" />
+</head>
+<body>
+  <div id="A01025119"></div>
+  <script type="module" src="/src/class1/A01025119/webpage/boot.tsx"></script>
+</body>
+</html>
