+++ 502/vite-project/src/class2/A01025119/Button.tsx
@@ -1,0 +1,17 @@
+import React from 'react';
+import '../../class1/A01025119/webpage/main_design.css';
+
+interface ButtonProps {
+  label: string;
+  onClick: () => void;
+}
+
+const Button: React.FC<ButtonProps> = ({ label, onClick }) => {
+  return (
+    <button className="main-button" onClick={onClick}>
+      {label}
+    </button>
+  );
+};
+
+export default Button;
