--- 502/vite-project/src/class2/A01025119/Tarea_2/Button.tsx
@@ -1,29 +1,0 @@
-import React from 'react';
-
-interface ButtonProps {
-  label: string;
-  onClick: () => void;
-}
-
-const Button: React.FC<ButtonProps> = ({ label, onClick }) => {
-  return (
-    <button onClick={onClick} style={styles.button}>
-      {label}
-    </button>
-  );
-};
-
-const styles = {
-  button: {
-    width: '100%',
-    padding: '10px',
-    backgroundColor: '#4CAF50',
-    color: 'white',
-    border: 'none',
-    borderRadius: '4px',
-    fontSize: '16px',
-    cursor: 'pointer',
-  } as React.CSSProperties,
-};
-
-export default Button;
