+++ 502/vite-project/src/class3/A01025119/Login_page.tsx
@@ -1,7 +1,7 @@
-import React, { useReducer } from 'react';
+import { useReducer } from 'react';
 import { useNavigate } from 'react-router-dom';
-import InputField from '../../class2/A01025119/Tarea_2/Input.tsx';
-import Button from '../../class2/A01025119/Tarea_2/Button.tsx';
+import InputField from '../../class2/A01025119/Input.tsx';
+import Button from '../../class2/A01025119/Button.tsx';
 
 type State = {
   username: string;
